+++ apps/docs/src/pages/permit/creating-policies.mdx
@@ -31,7 +31,7 @@ Resources are defined using [Standard Schema](https://standardschema.dev/), whic
 
 ```ts
 import { z } from "zod";
-import type { Resources } from "@zap-studio/permit/types";
+import type { Resources } from "@zap-studio/permit";
 
 const resources = {
   post: z.object({
@@ -59,7 +59,7 @@ const resources = {
 
 ```ts
 import * as v from "valibot";
-import type { Resources } from "@zap-studio/permit/types";
+import type { Resources } from "@zap-studio/permit";
 
 const resources = {
   post: v.object({
@@ -74,7 +74,7 @@ const resources = {
 
 ```ts
 import { type } from "arktype";
-import type { Resources } from "@zap-studio/permit/types";
+import type { Resources } from "@zap-studio/permit";
 
 const resources = {
   post: type({
@@ -92,7 +92,7 @@ const resources = {
 Actions specify what operations are allowed on each resource. Define them as readonly arrays:
 
 ```ts
-import type { Actions } from "@zap-studio/permit/types";
+import type { Actions } from "@zap-studio/permit";
 
 const actions = {
   post: ["read", "write", "delete", "publish"],
@@ -230,7 +230,7 @@ Here's a complete example for an e-commerce application:
 ```ts
 import { z } from "zod";
 import { createPolicy, allow, when, or } from "@zap-studio/permit";
-import type { Resources, Actions } from "@zap-studio/permit/types";
+import type { Resources, Actions } from "@zap-studio/permit";
 
 // Define resources
 const resources = {
