+++ apps/docs/src/pages/retry/index.mdx
@@ -60,20 +60,20 @@ deno add jsr:@zap-studio/retry
 
 ## Runtime Support
 
-| Runtime | Minimum version |
-| --- | --- |
-| Node.js | 18.0.0 |
-| Bun | 1.0.0 |
-| Deno | 1.42 |
-| Cloudflare Workers | Any current release |
-| Browsers | Chrome/Edge 98, Firefox 97, Safari 15.4 |
+| Runtime            | Minimum version                         |
+| ------------------ | --------------------------------------- |
+| Node.js            | 18.0.0                                  |
+| Bun                | 1.0.0                                   |
+| Deno               | 1.42                                    |
+| Cloudflare Workers | Any current release                     |
+| Browsers           | Chrome/Edge 98, Firefox 97, Safari 15.4 |
 
 Cancellation relies on `AbortSignal.reason`, which sets the browser minimums above. Deno 1.42 is the first release that can install packages from JSR (`deno add jsr:@zap-studio/retry`).
 
 ## Quick Example
 
 ```ts
-import { ExponentialBackoff } from "@zap-studio/retry/exponential-backoff";
+import { ExponentialBackoff } from "@zap-studio/retry";
 
 const policy = new ExponentialBackoff({
   maxAttempts: 5,
