+++ packages/fetch/package.json
@@ -40,9 +40,8 @@
     ".": "./dist/index.js",
     "./constants": "./dist/constants.js",
     "./errors": "./dist/errors.js",
+    "./fetch": "./dist/fetch.js",
     "./headers": "./dist/headers.js",
-    "./internal": "./dist/internal.js",
-    "./methods": "./dist/methods.js",
     "./request": "./dist/request.js",
     "./types": "./dist/types.js",
     "./url": "./dist/url.js",
