+++ packages/permit/package.json
@@ -34,8 +34,10 @@
   "types": "./dist/index.d.ts",
   "exports": {
     ".": "./dist/index.js",
+    "./conditions": "./dist/conditions.js",
     "./errors": "./dist/errors.js",
     "./helpers": "./dist/helpers.js",
+    "./policy": "./dist/policy.js",
     "./types": "./dist/types.js",
     "./package.json": "./package.json"
   },
