+++ packages/permit/src/index.ts
@@ -1,18 +1,36 @@
 /**
- * Policy composition and evaluation utilities.
+ * Public entrypoint for the permit package.
+ *
+ * Re-exports the full public API. Every symbol is also available from a
+ * dedicated subpath (`@zap-studio/permit/policy`,
+ * `@zap-studio/permit/conditions`, ...) for consumers who prefer granular
+ * imports. All exports are side-effect free and tree-shakeable.
  *
  * @module @zap-studio/permit
  */
 
-import type { StandardSchemaV1 } from "@zap-studio/validation";
-import { createStandardValidator } from "@zap-studio/validation";
-
-import { PolicyError } from "./errors.js";
-import type {
+export {
+  allow,
+  and,
+  collectInheritedRoles,
+  deny,
+  has,
+  hasRole,
+  not,
+  or,
+  when,
+} from "./conditions.js";
+export { PolicyError } from "./errors.js";
+export { assertNever } from "./helpers.js";
+export { createPolicy, mergePolicies, mergePoliciesAny } from "./policy.js";
+export type {
+  ActionPolicyMap,
   Actions,
   ConditionFn,
   Context,
+  Decision,
   InferAction,
+  InferPermission,
   InferResource,
   PermitConfig,
   Policy,
@@ -20,550 +38,5 @@ import type {
   Resources,
   Role,
   RoleHierarchy,
+  Rules,
 } from "./types.js";
-
-/**
- * Returns a policy function that always allows the action.
- *
- * @example
- * ```ts
- * const policy = createPolicy({
- *   resources,
- *   actions,
- *   rules: {
- *     post: {
- *       read: allow(), // Always allow reading posts
- *     },
- *   },
- * });
- * ```
- */
-export const allow =
-  <
-    TContext extends Context,
-    TAction extends string = string,
-    TResource = unknown,
-  >(): PolicyFn<TContext, TAction, TResource> =>
-  () =>
-    "allow";
-
-/**
- * Returns a policy function that always denies the action.
- *
- * @example
- * ```ts
- * const policy = createPolicy({
- *   resources,
- *   actions,
- *   rules: {
- *     post: {
- *       delete: deny(), // Never allow deleting posts
- *     },
- *   },
- * });
- * ```
- */
-export const deny =
-  <
-    TContext extends Context,
-    TAction extends string = string,
-    TResource = unknown,
-  >(): PolicyFn<TContext, TAction, TResource> =>
-  () =>
-    "deny";
-
-/**
- * Returns a policy function that allows or denies based on a condition.
- *
- * @example
- * ```ts
- * const policy = createPolicy({
- *   resources,
- *   actions,
- *   rules: {
- *     post: {
- *       write: when((ctx, action, resource) => ctx.user.id === resource.authorId),
- *     },
- *   },
- * });
- * ```
- */
-export const when =
-  <
-    TContext extends Context,
-    TAction extends string = string,
-    TResource = unknown,
-  >(
-    condition: ConditionFn<TContext, TAction, TResource>
-  ): PolicyFn<TContext, TAction, TResource> =>
-  (context, action, resource) =>
-    condition(context, action, resource) ? "allow" : "deny";
-
-/**
- * Returns a condition function that returns `true` if all conditions are met.
- *
- * @example
- * ```ts
- * const isOwnerAndPublished = and(
- *   (ctx, action, resource) => ctx.user.id === resource.authorId,
- *   (ctx, action, resource) => resource.status === "published"
- * );
- *
- * rules: {
- *   post: {
- *     delete: when(isOwnerAndPublished),
- *   },
- * }
- * ```
- */
-export const and =
-  <
-    TContext extends Context,
-    TAction extends string = string,
-    TResource = unknown,
-  >(
-    ...conditions: ConditionFn<TContext, TAction, TResource>[]
-  ): ConditionFn<TContext, TAction, TResource> =>
-  (context, action, resource) =>
-    conditions.every((condition) => condition(context, action, resource));
-
-/**
- * Returns a condition function that returns `true` if any condition is met.
- *
- * @example
- * ```ts
- * const isOwnerOrAdmin = or(
- *   (ctx, action, resource) => ctx.user.id === resource.authorId,
- *   (ctx, action, resource) => ctx.user.role === "admin"
- * );
- *
- * rules: {
- *   post: {
- *     write: when(isOwnerOrAdmin),
- *   },
- * }
- * ```
- */
-export const or =
-  <
-    TContext extends Context,
-    TAction extends string = string,
-    TResource = unknown,
-  >(
-    ...conditions: ConditionFn<TContext, TAction, TResource>[]
-  ): ConditionFn<TContext, TAction, TResource> =>
-  (context, action, resource) =>
-    conditions.some((condition) => condition(context, action, resource));
-
-/**
- * Returns a condition function that negates another condition.
- *
- * @example
- * ```ts
- * const isNotOwner = not((ctx, action, resource) => ctx.user.id === resource.authorId);
- *
- * rules: {
- *   post: {
- *     like: when(isNotOwner), // Can only like posts you don't own
- *   },
- * }
- * ```
- */
-export const not =
-  <
-    TContext extends Context,
-    TAction extends string = string,
-    TResource = unknown,
-  >(
-    condition: ConditionFn<TContext, TAction, TResource>
-  ): ConditionFn<TContext, TAction, TResource> =>
-  (context, action, resource) =>
-    !condition(context, action, resource);
-
-/**
- * Returns a condition function that checks if a context property equals a value.
- *
- * @example
- * ```ts
- * rules: {
- *   post: {
- *     write: when(has("role", "admin")), // Only admins can write
- *   },
- * }
- * ```
- */
-export const has =
-  <TContext extends Context, K extends keyof TContext>(
-    key: K,
-    value: TContext[K]
-  ): ConditionFn<TContext> =>
-  (context) =>
-    context[key] === value;
-
-/**
- * Collects all roles including inherited ones from a role hierarchy.
- *
- * @example
- * ```ts
- * type Role = "guest" | "user" | "admin";
- *
- * const hierarchy: RoleHierarchy<Role> = {
- *   guest: [],
- *   user: ["guest"],
- *   admin: ["user"],
- * };
- *
- * collectInheritedRoles(["admin"], hierarchy);
- * // Returns: Set { "admin", "user", "guest" }
- * ```
- */
-export const collectInheritedRoles = <TRole extends Role = Role>(
-  roles: TRole[],
-  hierarchy: RoleHierarchy<TRole>
-): Set<TRole> => {
-  const inherited = new Set<TRole>();
-
-  const add = (role: TRole): void => {
-    if (inherited.has(role)) {
-      return;
-    }
-
-    inherited.add(role);
-    const parents = hierarchy[role] ?? [];
-    for (const parent of parents) {
-      add(parent);
-    }
-  };
-
-  for (const role of roles) {
-    add(role);
-  }
-  return inherited;
-};
-
-/**
- * Call signatures for {@link hasRole}, preserving the with/without hierarchy overloads.
- */
-interface HasRoleFn {
-  <
-    TContext extends { role: Role | Role[] },
-    TAction extends string = string,
-    TResource = unknown,
-  >(
-    role: Role
-  ): ConditionFn<TContext, TAction, TResource>;
-  <
-    TContext extends { role: TRole | TRole[] },
-    TAction extends string = string,
-    TResource = unknown,
-    TRole extends Role = Role,
-  >(
-    role: TRole,
-    hierarchy: RoleHierarchy<TRole>
-  ): ConditionFn<TContext, TAction, TResource>;
-}
-
-/**
- * Returns a condition function that checks if the user has a specific role.
- * Supports role hierarchy for inherited permissions.
- *
- * @example
- * ```ts
- * // Without hierarchy
- * rules: {
- *   post: {
- *     delete: when(hasRole("admin")),
- *   },
- * }
- *
- * // With hierarchy
- * const hierarchy = {
- *   guest: [],
- *   user: ["guest"],
- *   admin: ["user"],
- * };
- *
- * rules: {
- *   post: {
- *     read: when(hasRole("guest", hierarchy)), // Admins and users can also read
- *   },
- * }
- * ```
- */
-export const hasRole: HasRoleFn =
-  (
-    role: Role,
-    hierarchy?: RoleHierarchy
-  ): ConditionFn<{ role: Role | Role[] }> =>
-  (context) => {
-    const userRoles = Array.isArray(context.role)
-      ? context.role
-      : [context.role];
-
-    if (hierarchy === undefined) {
-      return userRoles.includes(role);
-    }
-
-    const inherited = collectInheritedRoles(userRoles, hierarchy);
-    return inherited.has(role);
-  };
-
-/**
- * Splits a typed `resource:action` permission string into its parts.
- * Returns `null` when the string is malformed (missing/empty part or extra segments).
- */
-const parsePermission = <
-  TResources extends Resources,
-  TActions extends Actions<TResources>,
-  K extends keyof TResources & keyof TActions,
->(
-  permission: `${K & string}:${InferAction<TActions, K> & string}`
-): { action: InferAction<TActions, K>; resourceType: K } | null => {
-  const [resourceTypeValue, actionValue, ...rest] = permission.split(":");
-  if (
-    resourceTypeValue === undefined ||
-    resourceTypeValue.length === 0 ||
-    actionValue === undefined ||
-    actionValue.length === 0 ||
-    rest.length > 0
-  ) {
-    return null;
-  }
-
-  return {
-    action: actionValue,
-    // oxlint-disable-next-line typescript/no-unsafe-type-assertion -- Parsed permission strings are constrained by the typed permission template.
-    resourceType: resourceTypeValue as K,
-  };
-};
-
-/**
- * Creates a type-safe policy from resource schemas, actions, and rules.
- *
- * @example
- * ```ts
- * import { z } from "zod";
- * import { createPolicy, allow, deny, when } from "@zap-studio/permit";
- * import type { Resources, Actions } from "@zap-studio/permit/types";
- *
- * // Define resource schemas
- * const resources = {
- *   post: z.object({
- *     id: z.string(),
- *     authorId: z.string(),
- *     visibility: z.enum(["public", "private"]),
- *   }),
- *   comment: z.object({
- *     id: z.string(),
- *     postId: z.string(),
- *     authorId: z.string(),
- *   }),
- * } satisfies Resources;
- *
- * // Define actions per resource
- * const actions = {
- *   post: ["read", "write", "delete"],
- *   comment: ["read", "write"],
- * } as const satisfies Actions<typeof resources>;
- *
- * // Define context type
- * type AppContext = { user: { id: string; role: string } };
- *
- * // Create the policy
- * const policy = createPolicy<AppContext>({
- *   resources,
- *   actions,
- *   rules: {
- *     post: {
- *       read: when((ctx, action, resource) => resource.visibility === "public"),
- *       write: when((ctx, action, resource) => ctx.user.id === resource.authorId),
- *       delete: deny(),
- *     },
- *     comment: {
- *       read: allow(),
- *       write: when((ctx, action, resource) => ctx.user.id === resource.authorId),
- *     },
- *   },
- * });
- *
- * // Check permissions
- * const post = { id: "1", authorId: "user-1", visibility: "public" as const };
- * await policy.can(ctx, "post:read", post); // true
- * await policy.can(ctx, "post:write", post); // depends on ctx.user.id
- * ```
- */
-export const createPolicy = <
-  TContext extends Context,
-  TResources extends Resources = Resources,
-  TActions extends Actions<TResources> = Actions<TResources>,
->(
-  config: PermitConfig<TContext, TResources, TActions>
-): Policy<TContext, TResources, TActions> => {
-  const { rules, resources, actions } = config;
-  const validators = new Map<
-    keyof TResources,
-    (input: unknown) => Promise<StandardSchemaV1.Result<unknown>>
-  >();
-
-  const getValidatedResource = async <K extends keyof TResources>(
-    resourceType: K,
-    resource: InferResource<TResources, K>
-  ): Promise<InferResource<TResources, K> | null> => {
-    const validator = validators.get(resourceType);
-    if (validator === undefined) {
-      return null;
-    }
-    try {
-      const result = await validator(resource);
-      if (result.issues) {
-        return null;
-      }
-      return result.value;
-    } catch (error) {
-      console.warn(
-        `Resource validation failed for ${String(resourceType)}: ${String(error)}`
-      );
-      return null;
-    }
-  };
-
-  const hasAllowedAction = <K extends keyof TResources & keyof TActions>(
-    resourceType: K,
-    action: InferAction<TActions, K>
-  ): boolean => actions[resourceType]?.includes(action) ?? false;
-
-  const evaluatePolicy = <K extends keyof TResources & keyof TActions>(
-    context: TContext,
-    resourceType: K,
-    action: InferAction<TActions, K>,
-    resource: InferResource<TResources, K>
-  ): boolean => {
-    const policyFn = rules[resourceType]?.[action];
-    if (policyFn === undefined) {
-      return false;
-    }
-
-    try {
-      return policyFn(context, action, resource) === "allow";
-    } catch (error) {
-      console.warn(
-        `Policy evaluation error for ${String(resourceType)}.${action}: ${String(error)}`
-      );
-      return false;
-    }
-  };
-
-  for (const key of Object.keys(resources) as (keyof TResources)[]) {
-    const schema = resources[key];
-    if (schema === undefined) {
-      throw new PolicyError(`Missing schema for resource: ${String(key)}`);
-    }
-    const validator = createStandardValidator(schema);
-    validators.set(key, async (input: unknown) => await validator(input));
-  }
-
-  return {
-    async can<K extends keyof TResources & keyof TActions>(
-      context: TContext,
-      permission: `${K & string}:${InferAction<TActions, K> & string}`,
-      resource: InferResource<TResources, K>
-    ): Promise<boolean> {
-      const parsedPermission = parsePermission<TResources, TActions, K>(
-        permission
-      );
-      if (parsedPermission === null) {
-        return false;
-      }
-
-      const { action, resourceType } = parsedPermission;
-      if (!hasAllowedAction(resourceType, action)) {
-        return false;
-      }
-
-      const validatedResource = await getValidatedResource(
-        resourceType,
-        resource
-      );
-      if (validatedResource === null) {
-        return false;
-      }
-
-      return evaluatePolicy(context, resourceType, action, validatedResource);
-    },
-  };
-};
-
-const mergePoliciesWithStrategy = <
-  TContext extends Context,
-  TResources extends Resources = Resources,
-  TActions extends Actions<TResources> = Actions<TResources>,
->(
-  policies: Policy<TContext, TResources, TActions>[],
-  strategy: "allow-overrides" | "deny-overrides"
-): Policy<TContext, TResources, TActions> => ({
-  async can<K extends keyof TResources & keyof TActions>(
-    context: TContext,
-    permission: `${K & string}:${InferAction<TActions, K> & string}`,
-    resource: InferResource<TResources, K>
-  ): Promise<boolean> {
-    if (policies.length === 0) {
-      return false;
-    }
-    for (const policy of policies) {
-      // oxlint-disable-next-line no-await-in-loop -- Policies must evaluate sequentially to preserve short-circuit semantics.
-      const allowed = await policy.can(context, permission, resource);
-
-      if (strategy === "allow-overrides" && allowed) {
-        return true;
-      }
-      if (strategy === "deny-overrides" && !allowed) {
-        return false;
-      }
-    }
-    return strategy === "deny-overrides";
-  },
-});
-
-/**
- * Merges multiple policies into one using "deny-overrides" strategy.
- * If any policy denies, the merged policy denies. All must allow for the result to allow.
- *
- * @example
- * ```ts
- * const basePolicy = createPolicy({ ... });
- * const adminPolicy = createPolicy({ ... });
- *
- * const merged = mergePolicies(basePolicy, adminPolicy);
- * // Both policies must allow for the action to be permitted
- * ```
- */
-export const mergePolicies = <
-  TContext extends Context,
-  TResources extends Resources = Resources,
-  TActions extends Actions<TResources> = Actions<TResources>,
->(
-  ...policies: Policy<TContext, TResources, TActions>[]
-): Policy<TContext, TResources, TActions> =>
-  mergePoliciesWithStrategy(policies, "deny-overrides");
-
-/**
- * Merges multiple policies into one using "allow-overrides" strategy.
- * If any policy allows, the merged policy allows. All must deny for the result to deny.
- *
- * @example
- * ```ts
- * const guestPolicy = createPolicy({ ... });
- * const memberPolicy = createPolicy({ ... });
- *
- * const merged = mergePoliciesAny(guestPolicy, memberPolicy);
- * // If either policy allows, the action is permitted
- * ```
- */
-export const mergePoliciesAny = <
-  TContext extends Context,
-  TResources extends Resources = Resources,
-  TActions extends Actions<TResources> = Actions<TResources>,
->(
-  ...policies: Policy<TContext, TResources, TActions>[]
-): Policy<TContext, TResources, TActions> =>
-  mergePoliciesWithStrategy(policies, "allow-overrides");
