+++ packages/retry/package.json
@@ -31,12 +31,11 @@
   "exports": {
     ".": "./dist/index.js",
     "./abort": "./dist/abort.js",
+    "./base-policy": "./dist/base-policy.js",
     "./errors": "./dist/errors.js",
     "./exponential-backoff": "./dist/exponential-backoff.js",
     "./fixed-delay": "./dist/fixed-delay.js",
-    "./result-mode": "./dist/result-mode.js",
     "./sleep": "./dist/sleep.js",
-    "./throw-mode": "./dist/throw-mode.js",
     "./types": "./dist/types.js",
     "./package.json": "./package.json"
   },
