+++ packages/retry/src/exponential-backoff.ts
@@ -4,7 +4,7 @@
  * @module @zap-studio/retry/exponential-backoff
  */
 
-import { BaseRetryPolicy } from "./index.js";
+import { BaseRetryPolicy } from "./base-policy.js";
 import type { RetryDecision, RetryDecisionInput } from "./types.js";
 
 /**
