+++ packages/webhooks/package.json
@@ -37,8 +37,9 @@
   "exports": {
     ".": "./dist/index.js",
     "./errors": "./dist/errors.js",
-    "./types": "./dist/types/index.js",
-    "./utils": "./dist/utils/index.js",
+    "./router": "./dist/router.js",
+    "./types": "./dist/types.js",
+    "./utils": "./dist/utils.js",
     "./verify": "./dist/verify.js",
     "./package.json": "./package.json"
   },
